requirements.txt
celery==5.2.7
redis==4.4.4         # or amqp if you're using RabbitMQ
requests==2.28.1
python-dotenv==0.21.0
jinja2==3.1.2

# If you want to use imaplib, you can rely on the standard library
# or install imaplib2
# imaplib2==2.101

# For production-level usage, you might need google-api-python-client, etc.
# if using Gmail API for promotions label removal, e.g.:
# google-api-python-client==2.70.0
# google-auth==2.15.0
